Add start and end frames
Two changes were made:\n1. A random number of start and end frames were added to the start and ending hitting moments respectively, to prevent the rally from being overly cropped.\n2. A change was made to the input of the  script. Now, create a file with the names of all the jsons that require transformation and pass it in as an argument to script